+++ abagen/allen.py
@@ -8,9 +8,8 @@ from functools import reduce
 
 import numpy as np
 import pandas as pd
-from scipy.spatial.distance import cdist
 
-from . import datasets, io, probes, process, utils
+from . import datasets, io, probes, process, samples, utils
 
 import logging
 logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.INFO)
@@ -18,187 +17,7 @@ lgr = logging.getLogger('abagen')
 lgr_levels = dict(zip(range(3), [40, 20, 10]))
 
 
-def _assign_sample(sample, atlas, sample_info=None, atlas_info=None,
-                   tolerance=2):
-    """
-    Determines which parcel `sample` belongs to in `atlas`
-
-    Parameters
-    ----------
-    sample : (1, 3) array_like
-        Coordinates (ijk) of microarray sample in `atlas` space
-    atlas : niimg-like object
-        ROI image, where each ROI should be identified with a unique
-        integer ID
-    sample_info : pandas.DataFrame
-        A single row of an `annotation` file, corresponding to the given sample
-    atlas_info : pandas.DataFrame,
-        Dataframe containing information about the specified `atlas`. Must have
-        _at least_ columns 'id', 'hemisphere', and 'structure' containing
-        information mapping atlas IDs to hemisphere and broad structural class
-        (i.e., "cortex", "subcortex", "cerebellum"). Default: None
-    tolerance : int, optional
-        Distance (in mm) that a sample must be from a parcel for it to be
-        matched to that parcel. This is only considered if the sample is not
-        directly within a parcel. Default: 2
-
-    Returns
-    -------
-    label : int
-        Parcel label of `sample`
-    """
-
-    # pull relevant info from atlas
-    label_data = utils.check_img(atlas).get_data()
-
-    # expand provided coordinates to include those w/i `tolerance` of `coords`
-    # set a hard euclidean distance limit to account for different voxel sizes
-    coords = utils.expand_roi(sample, dilation=tolerance, return_array=True)
-    coords = coords[cdist(sample, coords).squeeze() < tolerance]
-
-    # grab non-zero labels for expanded coordinates
-    possible_labels = label_data[coords[:, 0], coords[:, 1], coords[:, 2]]
-    nz_labels = possible_labels[possible_labels.nonzero()]
-    labels, counts = np.unique(nz_labels, return_counts=True)
-
-    # if atlas_info and sample_info are provided, drop potential labels who
-    # don't match hemisphere or structural class defined in `sample_info`
-    if atlas_info is not None and sample_info is not None:
-        for old_label in labels:
-            new_label = _check_label(old_label, sample_info, atlas_info)
-            if old_label != new_label:
-                nz_labels[nz_labels == old_label] = new_label
-        labels, counts = np.unique(nz_labels[nz_labels.nonzero()],
-                                   return_counts=True)
-
-    # if there is still nothing in the vicinity, return 0
-    if labels.size == 0:
-        return 0
-    # if there is only one ROI in the vicinity, use that
-    elif labels.size == 1:
-        return labels[0]
-
-    # if more than one ROI in the vicinity, return the most frequent
-    indmax, = np.where(counts == counts.max())
-    if indmax.size == 1:
-        return labels[indmax[0]]
-
-    # if two or more parcels tied for neighboring frequency, use ROI
-    # with closest centroid to `coords`
-    centroids = utils.get_centroids(atlas, labels)
-    return labels[utils.closest_centroid(sample, centroids)]
-
-
-def _check_label(label, sample_info, atlas_info):
-    """
-    Checks that `label` defined by `sample_info` is coherent with `atlas_info`
-
-    Parameters
-    ----------
-    label : int
-        Tenative label for sample described by `sample_info`
-    sample_info : pandas.DataFrame
-        A single row of an `annotation` file, corresponding to the given sample
-    atlas_info : pandas.DataFrame,
-        Dataframe containing information about the atlas of interest. Must have
-        _at least_ columns 'id', 'hemisphere', and 'structure' containing
-        information mapping atlas IDs to hemisphere and broad structural class
-        (i.e., "cortex", "subcortex", "cerebellum"). Default: None
-
-    Returns
-    -------
-    label : int
-        New label for sample
-    """
-
-    cols = ['hemisphere', 'structure']
-
-    if label != 0:
-        sample_info = sample_info[cols]
-        atlas_info = atlas_info.loc[label][cols]
-        if not np.all(sample_info.values == atlas_info.values):
-            label = 0
-
-    return label
-
-
-def label_samples(annotation, atlas, atlas_info=None, tolerance=2):
-    """
-    Matches all microarray samples in `annotation` to parcels in `atlas`
-
-    Attempts to place each sample provided in `annotation` into a parcel in
-    `atlas`, where the latter is a 3D niimg-like object that contains parcels
-    each idnetified by a unique integer ID.
-
-    The function tries to best match samples in `annotation` to parcels defined
-    in `atlas` by:
-
-        1. Determining if the sample falls directly within a parcel,
-        2. Checking to see if there are nearby parcels by slowly expanding the
-           search space to include nearby voxels, up to a specified distance
-           (specified via the `tolerance` parameter),
-        3. Assigning the sample to the closest parcel if there are multiple
-           nearby parcels, where closest is determined by the parcel centroid.
-
-    If at any step a sample can be assigned to a parcel the matching process is
-    terminated. If there is still no parcel for a given sample after this
-    process the sample is provided a label of 0.
-
-    Parameters
-    ----------
-    annotation : (S, 13) pandas.DataFrame
-        Pre-loaded annotation information for a given AHBA donor
-    atlas : niimg-like object
-        A parcellation image in MNI space, where each parcel is identified by a
-        unique integer ID
-    atlas_info : pandas.DataFrame, optional
-        Filepath to or pre-loaded dataframe containing information about
-        `atlas`. Must have _at least_ columns 'id', 'hemisphere', and
-        'structure' containing information mapping atlas IDs to hemisphere and
-        broad structural class (i.e., "cortex", "subcortex", "cerebellum").
-        Default: None
-    tolerance : int, optional
-        Distance (in mm) that a sample must be from a parcel for it to be
-        matched to that parcel. This is only considered if the sample is not
-        directly within a parcel. Default: 2
-
-    Returns
-    -------
-    labels : (S, 1) pandas.DataFrame
-        Dataframe with parcel labels for each of `S` samples
-    """
-
-    # get annotation and atlas data
-    annotation = io.read_annotation(annotation)
-    atlas = utils.check_img(atlas)
-    label_data, affine = atlas.get_data(), atlas.affine
-
-    # load atlas_info, if provided
-    if atlas_info is not None:
-        atlas_info = utils.check_atlas_info(atlas, atlas_info)
-
-    # get ijk coordinates for microarray samples and find labels
-    g_ijk = utils.xyz_to_ijk(annotation[['mni_x', 'mni_y', 'mni_z']], affine)
-    labelled_samples = label_data[g_ijk[:, 0], g_ijk[:, 1], g_ijk[:, 2]]
-
-    # if sample coordinates aren't directly inside a parcel, increment radius
-    # around sample up to `tolerance` to try and find nearby parcels.
-    # if still no parcel, then ignore this sample
-    for idx in np.where(labelled_samples == 0)[0]:
-        label, tol = labelled_samples[idx], 1
-        while label == 0 and tol <= tolerance:
-            label = _assign_sample(g_ijk[[idx]], atlas,
-                                   sample_info=annotation.iloc[idx],
-                                   atlas_info=atlas_info,
-                                   tolerance=tol)
-            tol += 1
-        labelled_samples[idx] = label
-
-    return pd.DataFrame(labelled_samples, dtype=int,
-                        columns=['label'], index=annotation.index)
-
-
-def group_by_label(microarray, sample_labels, labels=None, metric='mean'):
+def groupby_label(microarray, sample_labels, labels=None, metric='mean'):
     """
     Averages expression data in `microarray` over samples with same label
 
@@ -492,26 +311,26 @@ def get_expression_data(atlas, atlas_info=None, *, exact=True,
                           index=np.append([0], all_labels))
     for subj in range(num_subj):
         # get rid of samples whose coordinates don't match ontological profile
-        annotation = process.drop_mismatch_samples(files['annotation'][subj],
+        annotation = samples.drop_mismatch_samples(files['annotation'][subj],
                                                    files['ontology'][subj],
                                                    corrected=corrected_mni)
 
         # subset representative probes + samples from microarray data
-        samples = microarray[subj].loc[annotation.index]
+        data = microarray[subj].loc[annotation.index]
 
         # assign samples to regions and aggregate samples w/i the same region
-        sample_labels = label_samples(annotation, atlas,
-                                      atlas_info=atlas_info,
-                                      tolerance=tolerance)
-        expression += [group_by_label(samples, sample_labels,
-                                      all_labels, metric=metric)]
+        labels = samples.label_samples(annotation, atlas,
+                                       atlas_info=atlas_info,
+                                       tolerance=tolerance)
+        expression += [groupby_label(data, labels,
+                                     all_labels, metric=metric)]
         lgr.info('{:>3} / {} samples matched to ROIs for donor #{}'
-                 .format(np.sum(np.asarray(sample_labels) != 0),
+                 .format(np.sum(np.asarray(labels) != 0),
                          len(annotation),
                          datasets.WELL_KNOWN_IDS.value_set('subj')[subj]))
 
         # get counts of samples collapsed into each ROI
-        labs, num = np.unique(sample_labels, return_counts=True)
+        labs, num = np.unique(labels, return_counts=True)
         counts.loc[labs, subj] = num
 
         # if we don't want to do exact matching then cache which parcels are
